+++ magda-scala-common/src/main/scala/au/csiro/data61/magda/model/Registry.scala
@@ -292,7 +292,7 @@ object Registry
       "dataset-format",
       "publishing",
       "spatial-coverage",
-      "dataset-access-control",
+      "access-control",
       "access",
       "provenance"
     )