+++ magda-typescript-common/src/authorization-api/model.ts
@@ -98,12 +98,7 @@ export interface UserToken {
  * You only need this when you try to figure out user characteristic (unknown) via Known information of a dataset.
  * i.e. set `input.user` to unknown then calculate residual rules via partial evaluation.
  */
-export interface DatasetAccessControlMetaData {
-    /**
-     * We do allow "archived" as defined in `publishing.schema.json`
-     * But we probably should avoid using it as there is no story behinds it.
-     */
-    publishingState: "draft" | "published" | "archived";
+export interface AccessControlMetaData {
     accessControl?: {
         ownerId?: string;
         orgUnitOwnerId?: string;